pyproject.toml
[tool.poetry]
name = "konverter"
version = "0.1.0"
description = "YAML templating for responsible users"
authors = ["Simon Westphahl <simon@westphahl.net>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/westphahl/konverter"
documentation = "https://github.com/westphahl/konverter/blob/master/README.md"
keywords = ["yaml", "template", "kubernetes"]

[tool.poetry.dependencies]
python = "^3.8"
cryptography = "^2.8"
"ruamel.yaml" = "^0.16.5"
click = "^7.0"
jinja2 = "^2.10.3"
mypy = "^0.761"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pytest = "^5.3.2"
flake8 = "^3.7.9"
flake8-bugbear = "^20.1.2"

[tool.poetry.scripts]
konverter = "konverter.__main__:cli"
konverter-vault = "konverter.vault:cli"

[tool.black]
line-length = 88
py36 = true
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
